Core: arm64 port — struct layout, pointer translation, UAE engine wiring
- modhcom fields long→uint32_t, short→uint16_t; fixes E_FNA (_msize at wrong offset)
- procid pointer fields →uint32_t (68k virtual addresses, not host pointers)
- sgstat fields →uint32_t/uint16_t; os9direntry_typ fdsect →uint32_t + static_assert
- Revert ulong to unsigned long (64-bit); codebase uses it to hold host pointers
- Fix pointer arithmetic truncation in modstuff/memstuff (uintptr_t, not ulong)
- Translate register args at icalls/fcalls boundary (TO68K/FROM68K)
- Wire UAE 68k engine; align regs_type with UAE regstruct + static-assert guard
- Fix strcpy self-overlap in os9exec_nt.c caught by macOS fortified libc

+++ Source/OS9exec_core/memstuff.c
@@ -721,7 +721,7 @@ os9err os9free( ushort pid, void* membase, ulong memsz )
         // try to free it in smaller pieces ...
         // NOTE: e.g. OS-9 "dir" is doing it this way !
         if (memsz > m->size) { // recursive call
-               err= os9free( pid, (void*)( (ulong)membase + m->size ), memsz - m->size );
+               err= os9free( pid, (void*)( (uintptr_t)membase + m->size ), memsz - m->size );
           if (!err) {
             release_memblock( pid, k ); // only release it, if all of them are fitting
             return 0; /* freed ok */
